Project1_code/google_images_scraping/requirements.txt
selenium
Pillow
requests